chore: neutral configuration loader
